Fix issue with azimuthal shear result indexing
In some cases, the code tried to put results in indices that were
outside of image. This occurred in the edge of the images in when
there are big enough gaps in the data that are jumped over.
Likely, data used previously in testing did not include such cases
and the error was not encountered.

To fix the issue, add a simple check to make sure that the indices
are inside image.

Also bump version.