Add thresholding function
Add a function for thresholding images that can be used to create
warnings.

Also change the way version is preserved. Now version is stored in
iris/_version.py and imported run-time and in setup.py. This allows
checking version in iris.__version__ and storing it in only one place.
Bump minor version.